Variable Initialization

+++ elysia.sh
@@ -1,6 +1,8 @@
 #/bin/sh
 
 ./build/elysia com $@
-nasm -felf64 ./output.asm
-ld -o output ./output.o ./runtime/elysiart_x86-64.o
-rm ./output.o
+if [ $? -eq 0 ]; then
+    nasm -felf64 ./output.asm
+    ld -o output ./output.o ./runtime/elysiart_x86-64.o
+    rm ./output.o
+fi
